Updated: Correct file structure for StyleSheet helper This is to follow the convention that exports should be in index.js file at the root of the folder, and to improve searchability of file related to StyleSheet
