+++ src/Card/Card.js
@@ -4,7 +4,7 @@ import * as React from 'react';
 import { View } from 'react-native';
 import { defaultTokens } from '@kiwicom/orbit-design-tokens';
 
-import StyleSheet from '../PlatformStyleSheet';
+import { StyleSheet } from '../PlatformStyleSheet';
 import Touchable from '../Button/Touchable';
 
 import type { StylePropType } from '../PlatformStyleSheet/StyleTypes';
