+++ src/CarrierLogo/CarrierLogo.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { View, Image, PixelRatio } from 'react-native';
 import { defaultTokens } from '@kiwicom/orbit-design-tokens';
-import StyleSheet from '../PlatformStyleSheet';
+import { StyleSheet } from '../PlatformStyleSheet';
 import type { Props } from './CarrierLogoTypes';
 import { SIZE_OPTIONS } from './consts';
 import { logoSizes } from './styles';
