+++ src/Checkbox/CheckboxIcon.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { View } from 'react-native';
 import { Icon } from '../Icon';
-import StyleSheet from '../PlatformStyleSheet';
+import { StyleSheet } from '../PlatformStyleSheet';
 import theme, { parsePxValue } from './styles';
 import type { CheckboxIconProps } from './CheckboxTypes';
