+++ src/LocalizedPrice/LocalizedPrice.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { defaultTokens } from '@kiwicom/orbit-design-tokens';
 import { Text } from '../Text';
-import StyleSheet from '../PlatformStyleSheet';
+import { StyleSheet } from '../PlatformStyleSheet';
 import type { LocalizedPriceProps } from './LocalizedPriceTypes';
 
 export default function LocalizedPrice({
