+++ src/PlaceCard/Image/CityImage.js
@@ -2,7 +2,7 @@
 
 import * as React from 'react';
 
-import StyleSheet from '../../PlatformStyleSheet';
+import { StyleSheet } from '../../PlatformStyleSheet';
 import NetworkImage from './NetworkImage';
 
 import type { StylePropType } from '../../PlatformStyleSheet/StyleTypes';
