+++ src/PlaceCard/Image/NetworkImage.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { Image } from 'react-native';
 
-import StyleSheet from '../../PlatformStyleSheet';
+import { StyleSheet } from '../../PlatformStyleSheet';
 
 import type { StylePropType } from '../../PlatformStyleSheet/StyleTypes';
