+++ src/Placeholder/shimmer/index.native.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { Image } from 'react-native';
 import Shimmer from 'react-native-shimmer';
-import StyleSheet from '../../PlatformStyleSheet';
+import { StyleSheet } from '../../PlatformStyleSheet';
 
 const ShimmeringLoader = ({ style, ...rest }: any) => (
   <Shimmer {...rest}>
