+++ src/PlatformStyleSheet/__flowtests__/Styles.flowtest.js
@@ -1,6 +1,6 @@
 // @flow
 
-import StyleSheet from '..';
+import { StyleSheet } from '..';
 import { type StyleObjectType, type StylePropType } from '../StyleTypes';
 
 /**
