+++ src/RadioButton/RadioButton.stories.js
@@ -6,7 +6,7 @@ import { select, boolean, text, withKnobs } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
 import { RadioButton } from '.';
 import { Text } from '../Text';
-import StyleSheet from '../PlatformStyleSheet';
+import { StyleSheet } from '../PlatformStyleSheet';
 
 const pressAction = action('radio-button-press');
 const bulletTypes = ['bullet', 'check'];
