+++ src/Text/Text.js
@@ -4,7 +4,7 @@ import * as React from 'react';
 import { Text as RNText, Platform } from 'react-native';
 import { defaultTokens } from '@kiwicom/orbit-design-tokens';
 
-import StyleSheet from '../PlatformStyleSheet';
+import { StyleSheet } from '../PlatformStyleSheet';
 import { createStylesGenerator } from '../utils';
 import {
   align as alignTypes,
